PN-454 Misaligned PubMedIDs on Patient Form
This is a very small bug fix where only on the last PubMed ID in the "Similar cases" section had an icon, resulting in minor misalignment.

The `leanpub` icon was created only once outside of the loop, hence it would only appear where it was last attached - the last line. Creating an icon for each PubMed ID line fixes it. Tested directly on browser using dev tools.